+++ app/src/App/DesktopApp.tsx
@@ -21,8 +21,8 @@ import { DeviceDetails } from '../pages/Devices/DeviceDetails'
 import { DevicesLanding } from '../pages/Devices/DevicesLanding'
 import { ProtocolRunDetails } from '../pages/Devices/ProtocolRunDetails'
 import { RobotSettings } from '../pages/Devices/RobotSettings'
-import { ProtocolsLanding } from '../pages/Protocols/ProtocolsLanding'
-import { ProtocolDetails } from '../pages/Protocols/ProtocolDetails'
+import { ProtocolsLanding } from '../pages/Desktop/Protocols/ProtocolsLanding'
+import { ProtocolDetails } from '../pages/Desktop/Protocols/ProtocolDetails'
 import { AppSettings } from '../pages/AppSettings'
 import { Labware } from '../pages/Labware'
 import { useSoftwareUpdatePoll } from './hooks'
@@ -32,7 +32,7 @@ import { IncompatibleModuleTakeover } from '../organisms/IncompatibleModule'
 import { OPENTRONS_USB } from '../redux/discovery'
 import { appShellRequestor } from '../redux/shell/remote'
 import { useRobot, useIsFlex } from '../organisms/Devices/hooks'
-import { ProtocolTimeline } from '../pages/Protocols/ProtocolDetails/ProtocolTimeline'
+import { ProtocolTimeline } from '../pages/Desktop/Protocols/ProtocolDetails/ProtocolTimeline'
 import { PortalRoot as ModalPortalRoot } from './portal'
 import { DesktopAppFallback } from './DesktopAppFallback'
