+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -32,7 +32,7 @@ import { ProtocolSetup } from '../pages/ProtocolSetup'
 import { RobotDashboard } from '../pages/RobotDashboard'
 import { RobotSettingsDashboard } from '../pages/RobotSettingsDashboard'
 import { ProtocolDashboard } from '../pages/ProtocolDashboard'
-import { ProtocolDetails } from '../pages/ProtocolDetails'
+import { ProtocolDetails } from '../pages/ODD/ProtocolDetails'
 import { QuickTransferFlow } from '../organisms/QuickTransferFlow'
 import { QuickTransferDashboard } from '../pages/QuickTransferDashboard'
 import { QuickTransferDetails } from '../pages/QuickTransferDetails'
