+++ app/src/organisms/OnDeviceDisplay/RobotDashboard/RecentRunProtocolCard.tsx
@@ -34,7 +34,7 @@ import {
   ANALYTICS_PROTOCOL_PROCEED_TO_RUN,
 } from '../../../redux/analytics'
 import { Skeleton } from '../../../atoms/Skeleton'
-import { useMissingProtocolHardware } from '../../../pages/Protocols/hooks'
+import { useMissingProtocolHardware } from '../../../pages/Desktop/Protocols/hooks'
 import { useCloneRun } from '../../ProtocolUpload/hooks'
 import { useRerunnableStatusText } from './hooks'
