+++ app/src/pages/Desktop/Protocols/ProtocolsLanding/__tests__/ProtocolsLanding.test.tsx
@@ -1,17 +1,17 @@
 import * as React from 'react'
 import { vi, it, describe } from 'vitest'
 import { screen } from '@testing-library/react'
-import { renderWithProviders } from '../../../../__testing-utils__'
+import { renderWithProviders } from '../../../../../__testing-utils__'
 
-import { ProtocolsEmptyState } from '../../../../organisms/ProtocolsLanding/ProtocolsEmptyState'
-import { getStoredProtocols } from '../../../../redux/protocol-storage'
-import { storedProtocolData } from '../../../../redux/protocol-storage/__fixtures__'
-import { ProtocolList } from '../../../../organisms/ProtocolsLanding/ProtocolList'
+import { ProtocolsEmptyState } from '../../../../../organisms/ProtocolsLanding/ProtocolsEmptyState'
+import { getStoredProtocols } from '../../../../../redux/protocol-storage'
+import { storedProtocolData } from '../../../../../redux/protocol-storage/__fixtures__'
+import { ProtocolList } from '../../../../../organisms/ProtocolsLanding/ProtocolList'
 import { ProtocolsLanding } from '..'
 
-vi.mock('../../../../redux/protocol-storage')
-vi.mock('../../../../organisms/ProtocolsLanding/ProtocolsEmptyState')
-vi.mock('../../../../organisms/ProtocolsLanding/ProtocolList')
+vi.mock('../../../../../redux/protocol-storage')
+vi.mock('../../../../../organisms/ProtocolsLanding/ProtocolsEmptyState')
+vi.mock('../../../../../organisms/ProtocolsLanding/ProtocolList')
 
 const render = () => {
   return renderWithProviders(<ProtocolsLanding />)[0]
