+++ app/src/pages/Desktop/Protocols/ProtocolsLanding/index.tsx
@@ -3,11 +3,11 @@ import { useDispatch, useSelector } from 'react-redux'
 import {
   fetchProtocols,
   getStoredProtocols,
-} from '../../../redux/protocol-storage'
-import { ProtocolsEmptyState } from '../../../organisms/ProtocolsLanding/ProtocolsEmptyState'
-import { ProtocolList } from '../../../organisms/ProtocolsLanding/ProtocolList'
+} from '../../../../redux/protocol-storage'
+import { ProtocolsEmptyState } from '../../../../organisms/ProtocolsLanding/ProtocolsEmptyState'
+import { ProtocolList } from '../../../../organisms/ProtocolsLanding/ProtocolList'
 
-import type { Dispatch, State } from '../../../redux/types'
+import type { Dispatch, State } from '../../../../redux/types'
 
 export function ProtocolsLanding(): JSX.Element {
   const dispatch = useDispatch<Dispatch>()
