+++ app/src/pages/ODD/ProtocolDetails/Parameters.tsx
@@ -16,8 +16,8 @@ import {
   TYPOGRAPHY,
   WRAP,
 } from '@opentrons/components'
-import { useToaster } from '../../organisms/ToasterOven'
-import { useRunTimeParameters } from '../Protocols/hooks'
+import { useToaster } from '../../../organisms/ToasterOven'
+import { useRunTimeParameters } from '../../../pages/Desktop/Protocols/hooks'
 import { EmptySection } from './EmptySection'
 import type { RunTimeParameter } from '@opentrons/shared-data'
