+++ app/src/pages/ODD/ProtocolDetails/__tests__/EmptySection.test.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { it, describe } from 'vitest'
 import { screen } from '@testing-library/react'
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 import { EmptySection } from '../EmptySection'
 
 const render = (props: React.ComponentProps<typeof EmptySection>) => {
