+++ app/src/pages/ODD/ProtocolDetails/__tests__/ProtocolDetails.test.tsx
@@ -4,7 +4,7 @@ import { fireEvent, screen, waitFor } from '@testing-library/react'
 import { when } from 'vitest-when'
 import { Route, MemoryRouter, Routes } from 'react-router-dom'
 import '@testing-library/jest-dom/vitest'
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 import { deleteProtocol, deleteRun, getProtocol } from '@opentrons/api-client'
 import {
   useCreateRunMutation,
@@ -12,15 +12,15 @@ import {
   useProtocolQuery,
   useProtocolAnalysisAsDocumentQuery,
 } from '@opentrons/react-api-client'
-import { i18n } from '../../../i18n'
-import { useHardwareStatusText } from '../../../organisms/OnDeviceDisplay/RobotDashboard/hooks'
-import { useOffsetCandidatesForAnalysis } from '../../../organisms/ApplyHistoricOffsets/hooks/useOffsetCandidatesForAnalysis'
+import { i18n } from '../../../../i18n'
+import { useHardwareStatusText } from '../../../../organisms/OnDeviceDisplay/RobotDashboard/hooks'
+import { useOffsetCandidatesForAnalysis } from '../../../../organisms/ApplyHistoricOffsets/hooks/useOffsetCandidatesForAnalysis'
 import {
   useMissingProtocolHardware,
   useRunTimeParameters,
-} from '../../Protocols/hooks'
-import { ProtocolSetupParameters } from '../../../organisms/ProtocolSetupParameters'
-import { formatTimeWithUtcLabel } from '../../../resources/runs'
+} from '../../../../pages/Desktop/Protocols/hooks'
+import { ProtocolSetupParameters } from '../../../../organisms/ProtocolSetupParameters'
+import { formatTimeWithUtcLabel } from '../../../../resources/runs'
 import { ProtocolDetails } from '..'
 import { Deck } from '../Deck'
 import { Hardware } from '../Hardware'
@@ -42,19 +42,19 @@ Object.defineProperty(window, 'IntersectionObserver', {
   configurable: true,
   value: IntersectionObserver,
 })
-vi.mock('../../../organisms/ProtocolSetupParameters')
+vi.mock('../../../../organisms/ProtocolSetupParameters')
 vi.mock('@opentrons/api-client')
 vi.mock('@opentrons/react-api-client')
-vi.mock('../../../organisms/OnDeviceDisplay/RobotDashboard/hooks')
+vi.mock('../../../../organisms/OnDeviceDisplay/RobotDashboard/hooks')
 vi.mock(
-  '../../../organisms/ApplyHistoricOffsets/hooks/useOffsetCandidatesForAnalysis'
+  '../../../../organisms/ApplyHistoricOffsets/hooks/useOffsetCandidatesForAnalysis'
 )
-vi.mock('../../Protocols/hooks')
+vi.mock('../../../../pages/Desktop/Protocols/hooks')
 vi.mock('../Deck')
 vi.mock('../Hardware')
 vi.mock('../Labware')
 vi.mock('../Parameters')
-vi.mock('../../../redux/config')
+vi.mock('../../../../redux/config')
 
 const MOCK_HOST_CONFIG = {} as HostConfig
 const mockCreateRun = vi.fn((id: string) => {})
